requirements.txt
astroid==2.2.5
autopep8==1.4.4
beautifulsoup4==4.6.0
certifi==2019.3.9
chardet==3.0.4
Django==2.2.3
django-debug-toolbar==2.0
django-modelcluster==4.4
django-taggit==0.24.0
django-treebeard==4.3
djangorestframework==3.9.4
draftjs-exporter==2.1.5
html5lib==1.0.1
idna==2.8
isort==4.3.18
lazy-object-proxy==1.3.1
mccabe==0.6.1
Pillow==5.4.1
pycodestyle==2.5.0
pylint==2.3.1
pytz==2019.1
requests==2.21.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
Unidecode==1.0.23
urllib3==1.24.3
wagtail==2.5.1
webencodings==0.5.1
Willow==1.1
wrapt==1.11.1
